% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianGonzalezWu}
\alias{bayesianGonzalezWu}
\title{bayesianGonzalezWu}
\usage{
bayesianGonzalezWu(stimuli, kappa = psiLogOdds(kappaObjective),
  kappaObjective = 0.5, tauStimuli = 1, tauCategory = 1,
  leftBoundaryObj = 0 - smallValue, rightBoundaryObj = 1 + smallValue,
  smallValue = 10^-10, leftBoundaryScaled = log(0 - leftBoundaryObj),
  rightBoundaryScaled = log(rightBoundaryObj - 1), responses = NULL,
  mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{kappa}{The location of the category}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{responses}{an optional vector of responses.
If responses are given, the return value is the logLikelihood of the responses given the parameters}

\item{leftBoundary}{The location of the posited (or fitted) psychological left-hand boundary of the screen. 0}

\item{rightBoundary}{The location of the posited (or fitted) psychological right-hand boundary of the screen. 1}
}
\value{
A vector the transformed stimuli
}
\description{
bayesianGonzalezWu
}
\examples{
bayesianGonzalezWu(1:100/100)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
