% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianHuttenlocherSpatialMemory}
\alias{bayesianHuttenlocherSpatialMemory}
\title{bayesianHuttenlocherSpatialMemory}
\usage{
bayesianHuttenlocherSpatialMemory(stimuli, kappa = 0, tauStimuli = 1,
  tauCategory = 1, boundary = 1, leftBoundary = -1 * boundary,
  rightBoundary = boundary, center = 0, responses = "none")
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}

\item{kappa}{The location of the categories (presumed symmetric on both sides around 0)}

\item{tauStimuli}{The precision of the stimulus traces: should be a single number}

\item{tauCategory}{The precision of the category distribution: should be a single number}

\item{responses}{an optional vector of responses. If responses are given, the return value is the logLikelihood of the responses given the parameters}

\item{boundaries}{The subject-specific location of the boundaries: may bear any relation to true stimuli, except that it 
should not leave real data outside the boundaries}
}
\value{
A vector the transformed stimuli, or the logLikelihood of them.
}
\description{
A simple package that assembles the symmetric model used by Huttenlocher and colleagues analyze spatial 
estimations from memory in one dimension
}
\examples{
bayesianHuttenlocherSpatialMemory(-99:100/100)
bayesianHuttenlocherSpatialMemory(-99:100/100, kappa=1, tauStimuli=2)
bayesianHuttenlocherSpatialMemory(1:100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
psiIdentity, multiCycleInverse
}
