% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianSpatialEstimationLandyCrawfordCorbin2017}
\alias{bayesianSpatialEstimationLandyCrawfordCorbin2017}
\title{bayesianSpatialEstimationLandyCrawfordCorbin2017}
\usage{
bayesianSpatialEstimationLandyCrawfordCorbin2017(stimuli, kappa = 0,
  tauStimuli = 1, tauCategory = 1, responses = "none")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{kappa}{The location of the category}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{responses}{an optional vector of responses. If responses are given, the return value is the}
}
\value{
A vector the transformed stimuli
}
\description{
A simple package that assembles the symmetric model used by Landy et al 2017 to analyze spatial 
estimations from memory in one dimension
}
\examples{
bayesianStevensPowerLaw(1:100)
bayesianStevensPowerLaw(1:100, kappa=1, tauStimuli=2)
bayesianStevensPowerLaw(1:100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
bayesianStevensPowerLaw
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
