% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianStevensPowerLaw}
\alias{bayesianStevensPowerLaw}
\title{bayesianStevensPowerLaw}
\usage{
bayesianStevensPowerLaw(stimuli, kappaObjective = 1,
  kappa = psiLog(kappaObjective), tauStimuli = 1, tauCategory = 1,
  responses = "none")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{kappaObjective}{An alternative specification giving kappa in objective units}

\item{kappa}{The location of the category}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{responses}{an optional vector of responses.
If responses are given, the return value is the logLikelihood of the responses given the parameters}
}
\value{
A vector the transformed stimuli
}
\description{
bayesianStevensPowerLaw
}
\examples{
bayesianStevensPowerLaw(1:100)
bayesianStevensPowerLaw(1:100, kappa=1, tauStimuli=2)
bayesianStevensPowerLaw(1:100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
