% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicPsychophysics.R
\name{classicStevensPowerLaw}
\alias{classicStevensPowerLaw}
\title{classicStevensPowerLaw}
\usage{
classicStevensPowerLaw(stimuli, exponent = 1, proportionality = 1,
  responses = NULL, tau = 1, mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{exponent}{the power to raise the stimulus to, equivalent to the ratio of taus in the reconsturcted version}

\item{proportionality}{the constant multiple on the stimulus value}

\item{responses}{an optional vector of responses.}

\item{tau}{if log-likelihood is to be calculated, we must decide the predicted variability of responses. This is captured in tau, the precision}

\item{mode}{a mode. can validly be "prediction", or "logLikelihoodOfResponses". If the latter, responses must be given, and tau must be given}
}
\value{
A vector the transformed stimuli
}
\description{
classicStevensPowerLaw
}
\examples{
classicStevensPowerLaw(1:100)
classicStevensPowerLaw(1:100, exponent=0.5, proportionality=2)
classicStevensPowerLaw(1:100, exponent=0.9, proportionality=2, responses=2*(1:100)^.9)
}
\seealso{
classicGonzalezWu, classicSpence
}
