% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{fitWarpedBayesModel}
\alias{fitWarpedBayesModel}
\title{fitWarpedBayesModel}
\usage{
fitWarpedBayesModel(model, stimuli, responses, responseGrid = NULL,
  initialPars = NULL, fixedPars = NULL, control = list(maxit = 5000,
  reltol = 1e-199), fit = TRUE, optimizing = "Objective Log Likelihood")
}
\arguments{
\item{model}{a model that has the general layout of the "bayesian..." models included in this package.}

\item{stimuli}{a vector of stimuli, in whatever raw format you like.}

\item{responses}{a vector of stimuli, in whatever raw format you like.}

\item{responseGrid}{a vector of possible responses, for discretizing the normal distribution}

\item{initialPars}{an initial set of any parameter values you expect optim to optimize over}

\item{fixedPars}{a fixed set of any parameter values you do not expect optim to optimize over}

\item{control}{Passed directly into optim's control parameter}

\item{fit}{If set to false, simply returns the function evaluation over the initial parameters. This is mostly useful for debugging}

\item{optimizing}{What criterion should be optimized?  Current valid values are "Objective RMSE" and "Subjective Log Likelihood". Subjective log likelihood is 
currently considered to be confusing, and is not recommended for the naive user.}
}
\value{
A tibble that contains one row for each stimulus/response pair, and includes 
several columns (see details for details)
}
\description{
A convenience function that packages several commonly popular moves that let a model do optimizaiotn.
}
\details{
The returned avalue includes two columns that are different on each line--the meanExpecation of the
fitted model, and one simulation sampled from the final parameters.  Several more columns pass through the results of hte
fit (value, counts, and convergence).  Finally, one column will be made per parameter. 
This format is a convenient one if you plan to attach your model fits (and predictions) to a stimulus tibble.
}
\examples{
a <- fitWarpedBayesModel(bayesianGonzalezWu, 
                         1:1000/1000, 
                         bayesianGonzalezWu(1:1000/1000, mode="simulation"), 
                         initialPars = c(kappa=1, tauStimuli=100, tauCategory=10))
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
