% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{guessingModel}
\alias{guessingModel}
\title{guessingModel}
\usage{
guessingModel(model, stimuli = NULL, responses = NULL,
  responseGrid = sort(unique(responses)), mode = "prediction",
  guessRate = 0, ...)
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{responses}{an optional vector of responses}

\item{responseGrid}{An optional vector of possible responses}

\item{mode}{What aspect should the function calculate? Legel choices include "prediction", "simulation", and "subjectiveLogLikelihood"
If responses are given, the return value is the logLikelihood of the responses given the parameters}

\item{...}{All parameters that should be passed into the main model}
}
\value{
A vector the transformed stimuli
}
\description{
guessingModel
}
\examples{
uniformGuessingModel(1:100/100)
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
