% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaBayes.R
\name{logOdds}
\alias{logOdds}
\title{Calculate the log of the odds for a set of proportions.}
\usage{
logOdds(proportions)
}
\arguments{
\item{proportions}{A vector of proportions, between 0 and 1}
}
\description{
Utility function to convert proportions into log odds (truncating proportions <.005 and >.995)
}
\examples{
logOdds()
}
