% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaBayes.R
\name{logOddsInverse}
\alias{logOddsInverse}
\title{Calculate the proportions for a vector of logs of the odds}
\usage{
logOddsInverse(x)
}
\arguments{
\item{proportions}{A vector of log odds, between negative and positive infinity}
}
\description{
Utility function to convert log odds into proportions. This function does not truncate, so it does not 
truly invert logOdds
}
\examples{
logOdds()
}
