% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{multiCycle}
\alias{multiCycle}
\title{multiCycle}
\usage{
multiCycle(x, ...)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf, or a list of vectors of stimuli}

\item{references}{A vector of values.  This should include -Inf and Inf, but these are implicitly added
if they are omitted}
}
\value{
A vector containing warped stimuli, or a list of vectors
}
\description{
Divides a mapping, of any sort (bounded, semi-bounded, or unbounded) into a finite number of distinct regions,
by creating 'references' inside the space. Functionally, multiCycle simply divides a vector of stimuli into 
A list of vectors of values based on their proportional distances between two references.  Usually, these
are then fed into a psi function which creates a multiCycle space containing multiple infinite unbounded regions
which are conceptually adjacent
}
\details{
For instance, a spatial region |----------------|
might be divided by a line of vertical symmetry into two regions
 |--------||--------|
 And each be turned int an unbounded Prelec region
 <--------><-------->
 with a command like -10:10 %>% multicycle(0) %>% psiPrelec
}
\examples{
multiCycle(-100:100, c(-100, 0, 100))
(-100:100/100) \%>\% multiCycle(-1, 0, 1) \%>\% psiLogOdds() \%>\% vanillaBayes() \%>\% psiLogOddsInverse() \%>\% multiCycleInverse(-1, 0, 1) # Implements Landy et al's model of one-dimensional spatial memory, with fixed boundaries
psiLinear(-10:10, shift=10, scaling=2)
-10:10 \%>\% psiLinear(shift=2, scaling=0.5)
}
\seealso{
psiIdentity, psiLogOdds, psiPrelec, psiLog, psiLinearInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
