% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiIdentityInverse}
\alias{psiIdentityInverse}
\title{psiIdentityInverse}
\usage{
psiIdentityInverse(x, ...)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}
}
\value{
a vector containing "unwarped" stimuli identical to the warped stimuli
}
\description{
This function is a generic transformation function that does no conversion at all. 
It is the inverse of psiIdentity.
It's really just here for testing things out.
It is equivalent to calling psiLinearInverse with default parameters
}
\examples{
psiIdentityInverse(-10:10)
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentity
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
