% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLinear}
\alias{psiLinear}
\title{psiLinear}
\usage{
psiLinear(stimuli, shift = 0, scaling = 1)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf, or a list of vectors of stimuli}

\item{shift}{A scalar by which to (positively) offset the scaled values, or a list or vector of scalars}

\item{scaling}{A scalar multiplier to the (unshifted) values, or a list or vector of scalars}
}
\value{
A vector containing warped stimuli, or a list of vectors
}
\description{
Creates mappings from -inf:+inf -> -inf:+inf, with a linear transformation mechanism
}
\examples{
psiLinear(c(0.1, 0.2, 0.3))
psiLinear(-10:10)
psiLinear(-10:10, shift=10, scaling=2)
-10:10 \%>\% psiLinear(shift=2, scaling=0.5)
}
\seealso{
psiIdentity, psiLogOdds, psiPrelec, psiLog, psiLinearInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
