% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLinearInverse}
\alias{psiLinearInverse}
\title{psiLinearInverse}
\usage{
psiLinearInverse(warpedStimuli, shift = 0, scaling = 1)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}
}
\value{
A vector containing unwarped stimuli
}
\description{
Creates mappings from -inf:+inf -> -inf:+inf, with a linear transformation mechanism
}
\examples{
psiLinear(c(0.1, 0.2, 0.3))
psiLinear(-10:10)
psiLinear(-10:10, shift=10, scaling=2)
-10:10 \%>\% psiLinear(shift=2, scaling=0.5)
}
\seealso{
psiIdentity, psiLogOdds, psiPrelec, psiLog, psiLinear
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
