% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLog}
\alias{psiLog}
\title{psiLog}
\usage{
psiLog(stimuli, smallValue = 10^-5)
}
\arguments{
\item{stimuli}{a vector of positive stimuli, between 0 and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses}
}
\value{
a vector containing "warped" stimuli
}
\description{
This function is a transformation function that takes the log...
It's literally just log(stimuli).
It's appropriate for mapping (0, +inf):-> (-inf, inf).
By design, it is robust to the inclusion of the occasional 0, which it maps to a very small value.
}
\examples{
psiLog(1:100)
1:100 \%>\% psiLog()
1:1000 \%>\% psiLog() \%>\% 
    vanillaBayes() \%>\% 
    psiLogInverse()  # Implements Stevens Power Law
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
