% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLogInverse}
\alias{psiLogInverse}
\title{psiLogInverse}
\usage{
psiLogInverse(warpedStimuli, smallValue = 10^-30)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}

\item{smallValue}{values smaller than this will be treated as this, to avoid anomalies near 0}
}
\value{
A vector containing unwarped stimuli
}
\description{
This function is a transformation function that inverts the log...
It's literally just exp(stimuli).
It's appropriate for mapping  (-inf, inf):-> (0, +inf)
}
\examples{
psiLogInverse(psiLog(1:100)) # returns 1:100
psiLog(1:100) \%>\% psiLogInverse()
1:1000 \%>\% psiLog() \%>\% vanillaBayes() \%>\% psiLogInverse()  # Implements Stevens Power Law
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
