% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLogOddsInverse}
\alias{psiLogOddsInverse}
\title{psiLogOddsInverse}
\usage{
psiLogOddsInverse(warpedStimuli, smallValue = 10^-30)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses. Accepted here to give the function and its inverse the same parameter set}
}
\value{
A vector containing unwarped stimuli
}
\description{
This function is a transformation function that inverts the log odds.
It's appropriate for mapping (-inf, inf) :->  (0, 1)
As in all functions in this package, the parameters are set so that applying the same parameters to the main function 
and the inverse yields an identity.
}
\examples{
psiLogOdds(1:100/100)
(0:1000/1000) \%>\% psiLogOdds() \%>\% 
    vanillaBayes() \%>\% 
    psiLogOddsInverse()  # Implements Gonzales & Wu, 1996
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
