% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiPrelecInverse}
\alias{psiPrelecInverse}
\title{psiPrelecInverse}
\usage{
psiPrelecInverse(warpedStimuli, smallValue = 10^-30)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses. Accepted here to give the function and its inverse the same parameter set}
}
\value{
A vector containing unwarped stimuli
}
\description{
This function is a transformation function inverts the Prelec transformation, ln(ln(1/s))
By design, it is robust to the inclusion of the occasional 0,
which it maps to a very small value, which can optionally be specified
It's appropriate for mapping (0, 1):-> (-inf, inf).
}
\examples{
psiPrelecInverse(-10:10)
(0:1000/1000) \%>\% psiPrelec() \%>\% 
    vanillaBayes() \%>\% 
    psiPrelecInverse()  # Implements Prelec, 1998
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLogInverse, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
