% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{uniCycleInverse}
\alias{uniCycleInverse}
\title{uniCycleInverse}
\usage{
uniCycleInverse(warpedStimuli, references = c(0))
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf, or a list of vectors of stimuli}

\item{references}{A vector of values.  This should include -Inf and Inf, but these are implicitly added
if they are omitted}
}
\value{
A vector containing warped stimuli, or a list of vectors
}
\description{
recombines a mapping, of any sort (bounded, semi-bounded, or unbounded) into single region,
by eliminating 'references' inside the space. Functionally, uniCycleInverse simply undoes a vector of stimuli into
}
\details{
For instance, a spatial region |----------------|
might be divided into three regions
 |=======||--------||=========|
 And the middle be turned into an unbounded Prelec region
 |=======|<-------->|=========|
 with a command like -10:10 %>% unicycle(0) %>% psiPrelec
 uniCycleInverse would then recombine these values into a single range.

KEY LIMITATION: Right now, uniCycleInverse can only 'handle' one layer of recursion. That is, you can't make calls like
 -10:10 %>% uniCycle %>% uniCycle %> uniCycleInverse %>% uniCycleInverse
 . Sorry.
}
\examples{
uniCycle(-99:100, c(-100, 0, 100))
plot(-99:100, unlist(uniCycle(-99:100, c(-100, -50, 0, 50, 100))))
(-99:100/100) \%>\% uniCycle(-1, 0, 1) \%>\% 
    psiLogOdds() \%>\% vanillaBayes() \%>\% 
    psiLogOddsInverse() \%>\% 
    uniCycleInverse(-1, 0, 1) # Implements Landy et al's model 
    # of one-dimensional spatial memory, with fixed boundaries
}
\seealso{
psiIdentity, psiLogOdds, psiPrelec, psiLog, psiLinearInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
