% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{uniformGuessingModel}
\alias{uniformGuessingModel}
\title{uniformGuessingModel}
\usage{
uniformGuessingModel(stimuli, responses = NULL,
  responseGrid = sort(unique(responses)), mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{responses}{an optional vector of responses}

\item{responseGrid}{An optional vector of possible responses}

\item{mode}{What aspect should the function calculate? Legel choices include "prediction", "simulation", and "likelihoodOfResponses"
If responses are given, the return value is the logLikelihood of the responses given the parameters}
}
\value{
A vector the transformed stimuli
}
\description{
uniformGuessingModel
}
\examples{
uniformGuessingModel(1:100/100)
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
