% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WarpedBayes-package.r
\docType{package}
\name{WarpedBayes}
\alias{WarpedBayes}
\alias{WarpedBayes-package}
\title{WarpedBayes. Psychophyisics = Space Warping + Bayes}
\description{
This package Implement several classic descriptive psychophysics laws as 
compositions of Bayes law and simple, principled transformations.
To see a full list of functions, do ls("package:WarpedBayes")
or help(package="WarpedBayes")
}
