% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianGonzalezWu}
\alias{bayesianGonzalezWu}
\title{bayesianGonzalezWu}
\usage{
bayesianGonzalezWu(stimuli, kappa = psiLogOdds(multiCycle(kappaObjective,
  references = c(leftBoundaryObjective, rightBoundaryObjective))),
  kappaObjective = 0.5, tauStimuli = 1, tauCategory = 1,
  leftBoundaryObjective = 0 - smallValue, rightBoundaryObjective = 1 +
  smallValue, rightBoundaryExpansion = log((rightBoundaryObjective -
  maxValue)/(maxValue - minValue)), leftBoundaryExpansion = log((minValue -
  leftBoundaryObjective)/(maxValue - minValue)), minValue = min(c(stimuli),
  na.rm = T), maxValue = max(c(stimuli), na.rm = T), smallValue = 10^-10,
  responses = NULL, mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between 0 and inf}

\item{kappa}{The location of the category in subjective space (from -inf to inf). Defaults to 0}

\item{kappaObjective}{An alternative specification for the kappa location, situated in objective measures.}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{leftBoundaryObjective}{The location (in objective units) of the posited (or fitted) psychological left-hand boundary of legal 
stimulus values. Defaults to a small amount less than 0}

\item{rightBoundaryObjective}{The location (in objective units) of the posited (or fitted) psychological right-hand boundary of legal 
stimulus values. Defaults to a small amount more than 1}

\item{rightBoundaryExpansion}{How much beyond the maximum stimulus/response value should the right boundary be expanded?}

\item{leftBoundaryExpansion}{How much beyond the minimium stimulus/response value should the left boundary be expanded?}

\item{minValue}{The lowest range value (in objective units).  Should be at least as small as the smallest stimulus and response.}

\item{maxValue}{The highest range value  (in objective units).  Should be at least as large as the largest stimulus and response.}

\item{smallValue}{a small amount, by which to default boundary expansion}

\item{responses}{an optional vector of responses}

\item{mode}{What aspect should the function calculate? Legel choices include "prediction", "simulation", and "logLikelihoodOfResponses"
If responses are given, the return value is the logLikelihood of the responses given the parameters}
}
\value{
A vector the transformed stimuli
}
\description{
bayesianGonzalezWu
}
\examples{
bayesianGonzalezWu(1:100/100)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
