% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{bayesianSpatialMemoryLandyCrawfordCorbin2017}
\alias{bayesianSpatialMemoryLandyCrawfordCorbin2017}
\title{bayesianSpatialMemoryLandyCrawfordCorbin2017}
\usage{
bayesianSpatialMemoryLandyCrawfordCorbin2017(stimuli,
  kappa = psiLogOdds(multiCycle(kappaObjective, references =
  c(leftBoundaryObjective, rightBoundaryObjective))), kappaObjective = 0.5,
  tauStimuli = 1, tauCategory = 1, leftBoundaryObjective = -1 -
  smallValue, rightBoundaryObjective = 1 + smallValue,
  rightBoundaryExpansion = log((rightBoundaryObjective - maxValue)/(maxValue -
  minValue)), leftBoundaryExpansion = log((minValue -
  leftBoundaryObjective)/(maxValue - minValue)), minValue = min(c(stimuli),
  na.rm = T), maxValue = max(c(stimuli), na.rm = T), smallValue = 10^-10,
  responses = NULL, center = 0, mode = "prediction")
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}

\item{kappa}{The location of the categories (presumed symmetric on both sides around the midline of the screen)}

\item{kappaObjective}{The location of the categories measured in objective units}

\item{tauStimuli}{The precision of the stimulus traces: should be a single number}

\item{tauCategory}{The precision of the category distribution: should be a single number}

\item{responses}{an optional vector of responses. If responses are given, the return value is the logLikelihood of the responses given the parameters}

\item{center}{The posited (or fitted) psychological center of the screen (in public units: should be near the true center)}

\item{boundary}{The subject-specific location of the boundaries: may bear any relation to true stimuli, except that it 
should not leave real data outside the boundaries}

\item{leftBoundary}{The location of the posited (or fitted) psychological left-hand boundary of the screen. Defaults to -1 * 'boundary'}

\item{rightBoundary}{The location of the posited (or fitted) psychological right-hand boundary of the screen. Defaults to 'boundary'}
}
\value{
A vector the transformed stimuli, or the logLikelihood of them.
}
\description{
A simple package that assembles the symmetric model used by Huttenlocher and colleagues analyze spatial 
estimations from memory in one dimension
}
\examples{
bayesianSpatialMemoryLandyCrawfordCorbin2017(-99:100/100)
bayesianSpatialMemoryLandyCrawfordCorbin2017(-99:100/100, kappa=1, tauStimuli=2)
bayesianSpatialMemoryLandyCrawfordCorbin2017(1:100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)
}
\seealso{
psiIdentity, multiCycleInverse
}
