% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructedPsychophysics.R
\name{fitWarpedBayesModel}
\alias{fitWarpedBayesModel}
\title{fitWarpedBayesModel}
\usage{
fitWarpedBayesModel(model, stimuli, responses, initialPars = NULL,
  fixedPars = NULL, control = list(maxit = 5000, reltol = 1e-119))
}
\arguments{
\item{model}{a model that has the general layout of the "bayesian..." models included in this package.}

\item{stimuli}{a vector of stimuli, in whatever raw format you like.}

\item{responses}{a vector of stimuli, in whatever raw format you like.}
}
\value{
A tibble that contains one row for each stimulus/response pair, and includes 
several columns (see details for details)
}
\description{
A convenience function that packages several commonly popular moves that let a model do optimizaiotn.
}
\details{
The returned avalue includes two columns that are different on each line--the meanExpecation of the
fitted model, and one simulation sampled from the final parameters.  Several more columns pass through the results of hte
fit (value, counts, and convergence).  Finally, one column will be made per parameter. 
This format is a convenient one if you plan to attach your model fits (and predictions) to a stimulus tibble.
}
\examples{
bayesianGonzalezWu(1:100/100)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2)
bayesianGonzalezWu(1:100/100, kappa=1, tauStimuli=2, responses=2*(1:100)^.9)#' @export
}
\seealso{
bayesianHuttenlocherSpatialMemory
}
