% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{multiCycle}
\alias{multiCycle}
\title{multiCycle}
\usage{
multiCycle(x, ...)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf, or a list of vectors of stimuli}

\item{references}{A vector of values.  This should include -Inf and Inf, but these are implicitly added
if they are omitted}
}
\value{
A vector containing warped stimuli, or a list of vectors
}
\description{
Divides a mapping, of any sort (bounded, semi-bounded, or unbounded) into a finite number of distinct regions,
by creating 'references' inside the space. Functionally, multiCycle simply divides a vector of stimuli into 
A list of vectors of values based on their proportional distances between two references.  Usually, these
are then fed into a psi function which creates a multiCycle space containing multiple infinite unbounded regions
which are conceptually adjacent
}
\details{
For instance, a spatial region |----------------|
might be divided by a line of vertical symmetry into two regions
 |--------||--------|
 And each be turned int an unbounded Prelec region
 <--------><-------->
 with a command like -10:10 %>% multicycle(0) %>% psiPrelec
}
\examples{
multiCycle(-99:100, c(-100, 0, 100))
(-99:100/100) \%>\% multiCycle(references= c(-1, 0, 1)) \%>\% psiLogOdds() \%>\% vanillaBayes() \%>\% psiLogOddsInverse() \%>\% multiCycleInverse(references=c(-1, 0, 1)) # Implements Landy et al's model of one-dimensional spatial memory, with fixed boundaries
plot(-99:100, unlist(multiCycle(-99:100, c(-100, -50, 0, 50, 100))))
plot(-99:100/100, (-99:100/100) \%>\% multiCycle(references= c(-1, 0, 1)) \%>\% psiLogOdds() \%>\% vanillaBayes(kappa=c(-.8, .8)) \%>\% psiLogOddsInverse() \%>\% multiCycleInverse(references=c(-1, 0, 1))-(-99:100/100), ylab="bias", xlab="stimulus");abline(0,0)
}
\seealso{
psiIdentity, multiCycleInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
