% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiIdentity}
\alias{psiIdentity}
\title{psiIdentity}
\usage{
psiIdentity(stimuli, smallValue = 10^-30)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses}
}
\value{
a vector containing "warped" stimuli identical to the original stimuli
}
\description{
This function is a generic transformation function that does no conversion at all. 
It's really just here for testing things out.
It is equivalent to calling psiLinear with default parameters
}
\examples{
psiIdentity(-10:10)
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
