% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiIdentityInverse}
\alias{psiIdentityInverse}
\title{psiIdentityInverse}
\usage{
psiIdentityInverse(warpedStimuli, smallValue = 10^-5)
}
\arguments{
\item{warpedStimuli}{a vector of stimuli, between -inf and inf}

\item{smallValue}{A value to set nominal 0's to, to avoid errors in plotting simulated data, or aberrant responses}
}
\value{
a vector containing "unwarped" stimuli identical to the warped stimuli
}
\description{
This function is a generic transformation function that does no conversion at all. 
It is the inverse of psiIdentity.
It's really just here for testing things out.
It is equivalent to calling psiLinearInverse with default parameters
}
\examples{
psiIdentityInverse(-10:10)
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentity
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
