% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLinear}
\alias{psiLinear}
\title{psiLinear}
\usage{
psiLinear(stimuli, shift = 0, scaling = 1)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}
}
\value{
a vector containing warped stimuli
}
\description{
Creates mappings from -inf:+inf -> -inf:+inf, with a linear transformation mechanism
}
\examples{
psiLinear(c(0.1, 0.2, 0.3))
psiLinear(-10:10)
psiLinear(-10:10, shift=10, scaling=2)
}
\seealso{
psiIdentity, psiLogOdds, psiPrelec, psiLog
}
\keyword{perceptual}
\keyword{psi,}
\keyword{tranformations}
