% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiFunctions.R
\name{psiLogInverse}
\alias{psiLogInverse}
\title{psiLogInverse}
\usage{
psiLogInverse(warpedStimuli, smallValue = 10^-5)
}
\arguments{
\item{stimuli}{a vector of positive stimuli, between 0 and inf}
}
\value{
a vector containing "warped" stimuli identical to the original stimuli
}
\description{
This function is a transformation function that inverts the log...
It's literally just exp(stimuli).
It's appropriate for mapping  (-inf, inf):-> (0, +inf)
}
\examples{
psiLogInverse(psiLog(1:100)) # returns 1:100
psiLog(1:100) \%>\% psiLogInverse()
1:1000 \%>\% psiLog() \%>\% vanillaBayes() \%>\% psiLogInverse()  # Implements Stevens Power Law
}
\seealso{
psiLinear, psiLogOdds, psiPrelec, psiLog, psiIdentityInverse
}
\keyword{perceptual}
\keyword{psi}
\keyword{tranformations}
