% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaBayes.R
\name{vanillaBayes}
\alias{vanillaBayes}
\title{vanillaBayes}
\usage{
vanillaBayes(stimuli, kappa = 0, tauStimuli = 1, tauCategory = 1,
  responses = "none")
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}

\item{kappa}{The location of the category}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{responses}{an optional vector of responses. If responses are given, the return value is the}
}
\value{
A vector containing mean stimulus locations
}
\description{
Performs vanilla bayes (single category) on a set of stimuli
}
\examples{
(0:1000/1000) \%>\% psiLogOdds() \%>\% vanillaBayes(kappa=5) \%>\% psiLogOddsInverse()  # Implements Gonzales & Wu, 1996
1:1000 \%>\% psiLog() \%>\% vanillaBayes() \%>\% psiLogInverse()  # Implements Stevens Power Law
}
\seealso{
vanillaBayes
}
\keyword{bayesianInference}
