% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VanillaBayes.R
\name{vanillaBayes}
\alias{vanillaBayes}
\title{vanillaBayes}
\usage{
vanillaBayes(stimuli, ...)
}
\arguments{
\item{stimuli}{a vector of stimuli, between -inf and inf}

\item{kappa}{The location of the category}

\item{tauStimuli}{The precision of the stimulus traces: may be a single number or a vector}

\item{tauCategory}{The precision of the category distribution}

\item{responses}{an optional vector of responses. If responses are given, the return value is the probability of these responses}
}
\value{
A vector containing mean stimulus locations
}
\description{
Performs vanilla bayes (single category) on a set of stimuli
}
\details{
If the kappa, tauStimuli, and tauCategory items are all more than length 1,
 and  are length 2 less than the number of bins, then we pad them by negative and positive infinity.
}
\examples{
(0:1000/1000) \%>\% psiLogOdds() \%>\% vanillaBayes(kappa=5) \%>\% psiLogOddsInverse()  # Implements Gonzales & Wu, 1996
1:1000 \%>\% psiLog() \%>\% vanillaBayes() \%>\% psiLogInverse()  # Implements Stevens Power Law
plot(-99:100/100, (-99:100/100) \%>\% multiCycle(references= c(-10, 0, 10)) \%>\% psiLogOdds() \%>\% vanillaBayes(kappa=c(-1, 1), tauStimuli=10) \%>\% psiLogOddsInverse() \%>\% multiCycleInverse(references=c(-10, 0, 10))-(-99:100/100), ylab="bias", xlab="stimulus");abline(0,0)
}
\seealso{
vanillaBayes
}
\keyword{bayesianInference}
